Tema06 - SQLAlchemy/Clase16_10-11-2025/api-libros-sqlalchemy/requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1